requirements.txt
astroid==2.3.2
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
geographiclib==1.50
geopy==1.20.0
get==2019.4.13
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-speech==1.2.0
google-cloud-texttospeech==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.24.3
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mlrose==1.3.0
numpy==1.17.3
post==2019.4.13
pprint==0.1
protobuf==3.10.0
public==2019.4.13
pyasn1==0.4.7
pyasn1-modules==0.2.7
pylint==2.4.3
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
typed-ast==1.4.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
